calorie_estimator/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.3,<2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
streamlit>=1.30.0
requests>=2.30.0
matplotlib>=3.7.1
opencv-contrib-python>=4.8.0
streamlit-webrtc>=0.47.0
python-dotenv>=1.0.0 